requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.5
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap3==12.1.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.2
entrypoints==0.3
Faker==4.0.3
gunicorn==20.0.4
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Markdown==3.3.3
MarkupSafe==1.1.1
misaka==2.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psycopg2==2.8.6
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.1.0
